+++ src/common/constant.ts
@@ -1,0 +1,1 @@
+export const LANG_ID = "language-hsp3";